initial poosh
very basic and bad live checker for streamers. if live it will initiate a linux (ubuntu) command to download the stream in question. meant to be used as a cronjob type deal.

streamlink is required. be sure to update the folder where you want the data saved in the subprocess.